# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Unify - Student Networking and Collaboration Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /freelance/services:
    get:
      operationId: listServices
      tags: [freelance]
      summary: List all freelance services
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Services
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceList"
    post:
      operationId: createService
      tags: [freelance]
      summary: Create a new freelance service listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceBody"
      responses:
        "200":
          description: Created service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /freelance/services/{id}:
    get:
      operationId: getService
      tags: [freelance]
      summary: Get a freelance service by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "404":
          description: Not found
    patch:
      operationId: updateService
      tags: [freelance]
      summary: Update a service you own
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceBody"
      responses:
        "200":
          description: Updated service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [freelance]
      summary: Delete a service you own
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required: [ok]
  /freelance/services/{id}/reviews:
    get:
      operationId: listServiceReviews
      tags: [freelance]
      summary: List reviews for a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewList"
    post:
      operationId: createServiceReview
      tags: [freelance]
      summary: Add a review to a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "200":
          description: Review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
components:
  schemas:
...
    ListingOwner:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        avatarColor:
          type: string
      required: [id, username, avatarColor]
...
    Review:
      type: object
      properties:
        id:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        reviewer:
          $ref: "#/components/schemas/ListingOwner"
      required: [id, rating, createdAt, reviewer]
    ReviewList:
      type: array
      items:
        $ref: "#/components/schemas/Review"
    CreateReviewBody:
      type: object
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
      required: [rating]
    Service:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        price:
          type: string
        category:
          type: string
        images:
          type: array
          items:
            type: string
        contactInfo:
          type: ["string", "null"]
        deliveryDays:
          type: integer
        createdAt:
          type: string
          format: date-time
        provider:
          $ref: "#/components/schemas/ListingOwner"
        avgRating:
          type: number
        reviewCount:
          type: integer
        isOwner:
          type: boolean
      required: [id, title, description, price, category, images, deliveryDays, createdAt, provider, avgRating, reviewCount, isOwner]
    ServiceList:
      type: array
      items:
        $ref: "#/components/schemas/Service"
    CreateServiceBody:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          minLength: 1
          maxLength: 2000
        price:
          type: number
          minimum: 0
        category:
          type: string
        images:
          type: array
          items:
            type: string
        contactInfo:
          type: string
        deliveryDays:
          type: integer
          minimum: 1
          maximum: 365
      required: [title, description, price, category]